refactor: 인증이 필요한 api와 필요하지 않은 api에 대한 instance 분리
